better image binding, fixes

+++ MTEngine/Engine/Core/SYS_Funct.mm
@@ -63,22 +63,22 @@ bool compare_str_num(char *str1, char *str2, u16 numChars)
 
 void FixFileNameSlashes(char *buf)
 {
-	u16 len = strlen(buf);
+	int len = strlen(buf);
 
 	// normalize path separators
-	for (u16 i = 0; i < len; i++)
+	for (int i = 0; i < len; i++)
 	{
 		if (buf[i] == '\\' || buf[i] == '/')
 			buf[i] = SYS_FILE_SYSTEM_PATH_SEPARATOR;
 	}
 	
 	// check if double path separators exist (fix for c:\dupa\/file.txt paths)
-	for (u16 i = 0; i < len-1; i++)
+	for (int i = 0; i < len-1; i++)
 	{
 		if (buf[i] == SYS_FILE_SYSTEM_PATH_SEPARATOR
 			&& buf[i+1] == SYS_FILE_SYSTEM_PATH_SEPARATOR)
 		{
-			for (u16 j = i; j < len-1; j++)
+			for (int j = i; j < len-1; j++)
 			{
 				buf[j] = buf[j+1];
 			}
@@ -90,8 +90,8 @@ void FixFileNameSlashes(char *buf)
 
 bool SYS_FileNameHasExtension(char *fileName, char *extension)
 {
-	u16 i = strlen(fileName) - 1;
-	u16 j = strlen(extension) - 1;
+	int i = strlen(fileName) - 1;
+	int j = strlen(extension) - 1;
 
 	// x.ext  - min 2 more
 	if (i < (j+2))
@@ -117,10 +117,10 @@ void SYS_RemoveFileNameExtension(char *fileName)
 	// warning! if the fileName is const it will crash...
 	// don't forget to not use extensions in the const char* filenames
 
-	u16 l = strlen(fileName);
+	int l = strlen(fileName);
 
 	bool isExt = false;
-	for (u16 z = 0; z < l; z++)
+	for (int z = 0; z < l; z++)
 	{
 		if (fileName[z] == '.')
 		{
@@ -131,7 +131,7 @@ void SYS_RemoveFileNameExtension(char *fileName)
 	if (isExt == false)
 		return;
 
-	u16 pos = l;
+	int pos = l;
 
 	int i = l-1;
